Added models and basic functionality to frontend. Forgot to change branch when I went to work on backend, so these are added too: Fixed services, controllers to use DTOs. Fixed mappers, added new interfaces. Got database working with H2 in memory.
However, data fetching doesn't work yet. Fix tomorrow.

+++ src/main/java/com/summer/flightsandseats/service/BookingService.java
@@ -1,5 +1,8 @@
 package com.summer.flightsandseats.service;
 
+import com.summer.flightsandseats.dto.BookingDTO;
+import com.summer.flightsandseats.mapper.BookingMapper;
+import com.summer.flightsandseats.mapper.FlightMapper;
 import com.summer.flightsandseats.model.Booking;
 import com.summer.flightsandseats.repository.BookingRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,19 +17,27 @@ public class BookingService {
 
     private final BookingRepository bookingRepository;
 
-    public List<Booking> getAllBookings() {
-        return bookingRepository.findAll();
+    public List<BookingDTO> getAllBookings() {
+        List<Booking> bookings = bookingRepository.findAll();
+        return BookingMapper.INSTANCE.toDTOList(bookings);
     }
 
-    public Optional<Booking> getBooking(Integer id) {
-        return bookingRepository.findById(id);
+    public BookingDTO getBooking(Integer id) {
+        Booking booking = bookingRepository.findById(id).orElse(null);
+        return booking != null ? BookingMapper.INSTANCE.toDto(booking) : null;
     }
 
-    public Booking saveBooking(Booking booking) {
-        return bookingRepository.save(booking);
+    public BookingDTO saveBooking(BookingDTO bookingDTO) {
+        Booking booking = BookingMapper.INSTANCE.toEntity(bookingDTO);
+        Booking savedBooking = bookingRepository.save(booking);
+        return BookingMapper.INSTANCE.toDto(savedBooking);
     }
 
-    public void deleteBooking(Integer id) {
-        bookingRepository.deleteById(id);
+    public boolean deleteBooking(Integer id) {
+        if (bookingRepository.existsById(id)) {
+            bookingRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
